split GROQ queries
